优化角度环转向

+++ Core/Bsp/Frame/TTS_Ex/tts_motor.h
@@ -67,9 +67,9 @@ typedef struct
 
 typedef struct
 {
-    int16_t X;
-    int16_t Y;
-    int16_t Z;
+    float X;
+    float Y;
+    float Z;
 } MotorAngle_t;
 
 typedef struct
